requirements.txt
streamlit==1.32.0
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
openai-whisper==20231117
ollama==0.1.6
requests==2.31.0
python-dotenv==1.0.1 